requirements.txt
aiohttp==3.8.1
aiohttp-retry==2.8.3
aiosignal==1.2.0
alembic==1.8.1
amqp==5.1.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asttokens==2.0.8
async-timeout==4.0.2
asyncssh==2.12.0
atpublic==3.1.1
attrs==22.1.0
autopage==0.5.1
backcall==0.2.0
billiard==3.6.4.0
birch==0.0.35
blis==0.9.1
cachetools==5.2.0
catalogue==2.0.8
catboost==1.0.6
celery==5.2.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cliff==4.0.0
cmaes==0.8.2
cmd2==2.4.2
colorama==0.4.5
colorlog==6.7.0
commonmark==0.9.1
confection==0.0.1
configobj==5.0.6
coverage==6.4.4
cryptography==37.0.4
cycler==0.11.0
cymem==2.0.6
debugpy==1.6.3
decorator==5.1.1
decore==0.0.1
dictdiffer==0.9.0
diskcache==5.4.0
distro==1.7.0
dpath==2.0.6
dulwich==0.20.45
dvc==2.24.0
dvc-data==0.4.0
dvc-gs==2.19.1
dvc-http==2.19.1
dvc-objects==0.2.0
dvc-render==0.0.9
dvc-task==0.1.2
dvclive==0.10.0
entrypoints==0.4
executing==0.10.0
fastai==2.7.9
fastcore==1.5.26
fastdownload==0.0.7
fastprogress==1.0.3
flatten-dict==0.4.2
flufl.lock==7.1
fonttools==4.37.1
frozenlist==1.3.1
FRUFS==1.0.2
fsspec==2022.8.2
funcy==1.17
future==0.18.2
gcsfs==2022.8.2
gitdb==4.0.9
GitPython==3.1.27
google-api-core==2.10.0
google-auth==2.11.0
google-auth-oauthlib==0.5.2
google-cloud-core==2.3.2
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
grandalf==0.6
graphviz==0.20.1
greenlet==1.1.3
hydra-core==1.2.0
icecream==2.1.3
idna==3.3
imbalanced-learn==0.9.1
importlib-metadata==4.12.0
iniconfig==1.1.1
ipykernel==6.15.1
ipython==8.4.0
ipywidgets==8.0.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jupyter-client==7.3.4
jupyter-core==4.11.1
jupyterlab-widgets==3.0.3
kiwisolver==1.4.4
kombu==5.2.4
langcodes==3.3.0
Mako==1.2.2
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
multidict==6.0.2
murmurhash==1.0.8
nanotime==0.5.2
nest-asyncio==1.5.5
networkx==2.8.6
numpy==1.23.2
oauthlib==3.2.0
omegaconf==2.2.3
optuna==3.0.0
packaging==21.3
pandas==1.4.3
parso==0.8.3
pathlib==1.0.1
pathspec==0.9.0
pathy==0.6.2
pbr==5.10.0
pdpipe==0.3.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
plotly==5.10.0
pluggy==1.0.0
preshed==3.0.7
prettytable==3.4.1
prompt-toolkit==3.0.30
protobuf==4.21.5
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.9.2
pydot==1.4.2
pygit2==1.10.1
Pygments==2.13.0
pygtrie==2.5.0
pyparsing==3.0.9
pyperclip==1.8.2
pytest==7.1.2
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
pyzmq==23.2.1
requests==2.28.1
requests-oauthlib==1.3.1
rich==12.5.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-learn==1.1.2
scipy==1.8.1
scmrepo==0.0.25
seaborn==0.12.0
shortuuid==1.0.9
shtab==1.5.5
six==1.16.0
skutil==0.0.18
smart-open==5.2.1
smmap==5.0.0
sortedcontainers==2.4.0
spacy==3.4.1
spacy-legacy==3.0.10
spacy-loggers==1.0.3
SQLAlchemy==1.4.40
srsly==2.4.4
stack-data==0.4.0
stevedore==4.0.0
strct==0.0.32
tabulate==0.8.10
tenacity==8.0.1
thinc==8.1.1
threadpoolctl==3.1.0
tomli==2.0.1
tomlkit==0.11.4
torch==1.12.1
torchvision==0.13.1
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
typer==0.4.2
typing_extensions==4.3.0
urllib3==1.26.12
vine==5.0.0
voluptuous==0.13.1
wasabi==0.10.1
wcwidth==0.2.5
widgetsnbextension==4.0.3
xgboost==1.6.2
yarl==1.8.1
zc.lockfile==2.0
zipp==3.8.1
